requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
drf-nested-routers==0.93.4
pip-autoremove==0.10.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
django-rest-knox==4.2.0
django-cors-headers==4.2.0
